fitPsf: ensure peak is in PSF area
If working with a CoaddPsf and the peak is just off the image, then
CachingPsf will give us a psf image at the wrong position (x0,y0 doesn't
correspond to the peak).  In that case, the assertion in _overlap() will
be hit because pbb is empty.  We check for this case explicitly.