Clarify tiny footprint limit
The tiny footprint limit cannot be reduced below 2 because the
"PSF dx" calculation involves shifting the PSF by one pixel to
the left and to the right (for a total of 2; and then we need
some pixels remaining, so the footprint must be larger than that).
Added comments to this effect, and enforced the limit in the
SourceDeblendConfig.