consolidate failed peak logic and downgrade warning
A peak is treated as "failed" if it has been explicitly marked
to be skipped or it doesn't have a HeavyFootprint.  These cases
are treated in the same code, so consolidated them so it's easier
to follow.

In the process, changed the warning to a debug message, to avoid
overwhelming the user: it is common for a peak to 'fail' in this
manner (usually due to having a 'tiny' Footprint, or the peak
being off the image), and the source is marked as skipped, so
we know what happened and there's no need to record this through
the logger (in the same way we don't log every exception from a
measurement algorithm).