identify non-deblended sources through nChild and mask
Sources that aren't deblended are already identified with the
deblend.skipped flag, but people tend not to use this when looking to
separate children and parents. Therefore, set deblend.nchild for sources
that aren't deblended so they are easily identifiable as parents (of
stillborn deblends).

Mask the non-deblended source footprints as NOT_DEBLENDED.

Both of these are accomplished in common code (new method skipParent) for
the two cases that we skip deblending a parent (big footprint, masked).