fitPsf: fix non-overlapping bounding boxes
Was getting an AssertionError from the _overlap internal function
because the bounding boxes didn't overlap at all. This was surprising
because the list of peaks should overlap by construction. On closer
inspection, found that at construction time the LOCAL bounding box was
checked, but the assertion uses the default frame (which is now PARENT).
Switched to using the default (PARENT) frame for construction for
consistency (LOCAL isn't used anywhere else in the deblender).